Use `maxSleepBetweenScans` value which is the `sleepBetweenScans` + the maximum variation. Remove `test-int` make target.
